Flytter kode for opphør fra LæremidlerBeregnYtelseSteg til LæremidlerBeregningService - Ønsker all logikk med å sette sammen beregningsresultat på et sted
